[refactor] #100 Participant 관련 dto 파일 폴더 변경
