+++ src/main/java/org/festimate/team/api/participant/dto/TypeRequest.java
@@ -1,4 +1,4 @@
-package org.festimate.team.api.festival.dto;
+package org.festimate.team.api.participant.dto;
 
 public record TypeRequest(
         Boolean q1, Boolean q2, Boolean q3, Boolean q4, Boolean q5
