+++ src/main/java/org/festimate/team/domain/participant/service/impl/ParticipantServiceImpl.java
@@ -4,8 +4,8 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.festimate.team.global.response.ResponseError;
 import org.festimate.team.global.exception.FestimateException;
-import org.festimate.team.api.festival.dto.TypeRequest;
-import org.festimate.team.api.festival.dto.TypeResponse;
+import org.festimate.team.api.participant.dto.TypeRequest;
+import org.festimate.team.api.participant.dto.TypeResponse;
 import org.festimate.team.domain.festival.entity.Festival;
 import org.festimate.team.api.participant.dto.ProfileRequest;
 import org.festimate.team.domain.participant.entity.Participant;
